Windows port for the TCP framework class
